Corretto errore esecuzione della ricerca dopo aver eseguito una add/change. Corretto problema visualizzazione delle risposte di una risorsa dopo operazione del.
git-svn-id: svn://openspcoop.org/gov4j/openspcoop2/branches/3.0/3.0.dev@13626 f7a281a6-e321-0410-a5c7-cb86b8957a31

+++ tools/web_interfaces/control_station/src/org/openspcoop2/web/ctrlstat/servlet/apc/AccordiServizioParteComuneCore.java
@@ -601,7 +601,7 @@ public class AccordiServizioParteComuneCore extends ControlStationCore {
 
 	}
 	
-	public List<org.openspcoop2.core.registry.ResourceResponse> accordiResourceResponseList(int idRisorsa, ISearch ricerca) throws DriverRegistroServiziException {
+	public List<org.openspcoop2.core.registry.ResourceResponse> accordiResourceResponseList(Long idRisorsa, ISearch ricerca) throws DriverRegistroServiziException {
 		Connection con = null;
 		String nomeMetodo = "accordiResourceResponseList";
 		DriverControlStationDB driver = null;
