+++ tools/web_interfaces/control_station/src/org/openspcoop2/web/ctrlstat/servlet/apc/AccordiServizioParteComuneResourcesRisposteAdd.java
@@ -228,7 +228,7 @@ public final class AccordiServizioParteComuneResourcesRisposteAdd extends Action
 				}
 			}
 			
-			List<ResourceResponse> lista = apcCore.accordiResourceResponseList(risorsa.getId().intValue(), ricerca);
+			List<ResourceResponse> lista = apcCore.accordiResourceResponseList(risorsa.getId(), ricerca);
 
 			apcHelper.prepareAccordiResourcesResponseList(ricerca, lista, id, as, tipoAccordo, nomeRisorsa);
