+++ tools/web_interfaces/control_station/src/org/openspcoop2/web/ctrlstat/servlet/apc/AccordiServizioParteComuneResourcesRisposteChange.java
@@ -279,7 +279,7 @@ public final class AccordiServizioParteComuneResourcesRisposteChange extends Act
 				}
 			}
 			
-			List<ResourceResponse> lista = apcCore.accordiResourceResponseList(risorsa.getId().intValue(), ricerca);
+			List<ResourceResponse> lista = apcCore.accordiResourceResponseList(risorsa.getId(), ricerca);
 
 			apcHelper.prepareAccordiResourcesResponseList(ricerca, lista, id, as, tipoAccordo, nomeRisorsa);
